Move typedefs to separate header
